Bookmarks: Fix issue with deletion of ancestor folders
The bookmarks reducers appear to have been intentionally using the
pre-update node state in order to find the selected folder's next
ancestor when a lower layer ancestor is deleted. Thus, the
speculative fix at
https://chromium-review.googlesource.com/c/chromium/src/+/6433201
was incorrect. This CL reverts the part of that change that updated
the reducers and updates existing tests such that they would have
caught the regression.

Bug: 412951110
Change-Id: I181910421bef1ff7bf5e1df729bd488d53bc0863
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6499062
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454148}