Keep split tabs intact when pinning/unpinning
Make SetTabsPinned move the entire split collection if a tab in that
split needs to be pinned/unpinned. Factored out code for sending out
notifications when multiple tabs are moving at once to a wrapper
function so the logic can be shared with other bulk move operations.
Also GetIndicesForCommand is changed so that if the context index is
part of a split and not selected, both tabs are returned.

Fixed: 410925548
Change-Id: I82e8f0a9783a107c9db97f2fa4870360c4886264
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6490566
Reviewed-by: Shibalik Mohapatra <shibalik@chromium.org>
Commit-Queue: Charles Meng <charlesmeng@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Auto-Submit: Charles Meng <charlesmeng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454170}