Treat compiled-in `additional_certs` as trust anchors when indicated
This CL adds a boolean `tls_trust_anchor` field to the root store proto,
for the purpose of indicating which `additional_certs` should be treated
as trust anchors. `additional_certs` marked as `tls_trust_anchor` are
added from the root store proto to the list of root certificates that is
loaded into TrustStoreChrome, alongside an associated trust anchor ID if
present. Functionally, this means that TAI intermediates, once they are
added to the root store with `tls_trust_anchor` set to true, will be
treated as trust anchors, when the root store is loaded from compiled-in
data. Expiry and X.509 constraints are enforced on these certificates
when the respective flags are set.

A follow-up CL will make the analogous change for root store data loaded
from the component updater.

Note that TrustStoreChrome does not yet do anything with the actual
Trust Anchor IDs themselves. They are dropped as ChromeRootStoreData is
constructed. A later change will store the trust anchor IDs in
TrustStoreChrome so that we can intersect server-provided TAIs with
Chrome's trusted TAIs and thereby decide which TAIs to advertise in the
TAI extension.

Bug: 396479339
Change-Id: I2f76478a4ebf40674d8c9751d9340f4b7709cea7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6481700
Commit-Queue: Emily Stark <estark@chromium.org>
Reviewed-by: Nick Harper <nharper@chromium.org>
Reviewed-by: Matt Mueller <mattm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454209}