Add a universal mojom error type for use with result<T,E>
This change adds a new universal type that can be used for most
result<T,E> impls, so that API designers do not need to reimplement
their own error types for their interface. It is very similar to
absl::Status types, but omits the kOk state, which should never be
used.

Note that this currently design does not allow easy propagation of
debug information. Users would have to declare their own error type
if they'd like to propagate their own debug information (e.g.: HRESULT).
Part of the reason this is not supported is because:

1. Mojo lack a side channel to easily stuff this additional data into.
2. The type system does not allow for easy extension of types.
3. Having a huge details struct where users have to figure out which
   fields to fill out can be cumbersome.

Additionally, because mojo does not have any centralized error
reporting, we've decided to implement a more ruidmentary error object.

If this is a major issue in the future, some future work could include:

1. Adding a new native mojo type which allows users to extend the
   error with their own specific details struct.
2. Add a payload to the result shadow-type.

Change-Id: If594a6e69fa3a094cc1a462fdced325e67728f84
Bug: 40841428
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6486488
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Fred Shih <ffred@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454211}