Publish events for sent/dropped attribution reports in DevTools
The report body is an object whose properties are not specified in the
protocol itself in order to avoid compatibility issues with respect to
the Attribution Reporting API's evolution, and because each report type
has a different set of fields.

The net/HTTP status details use the same naming scheme and format as
analogous types elsewhere in browser_protocol.pdl.

Bug: 40273989
Change-Id: If9aa2c18058e9f970e4f0a5f41771064df3e1f2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6494173
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Nan Lin <linnan@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454234}