Add argument to NetworkFetcher::PostRequestComplete
To allow for remote event logging, the updater needs access to the
'Cookie' header returned in POST request responses by the remote
logging server. The current `NetworkFetcher` interface does not
support this, so an additional argument is required in the type
signature of the `PostRequestComplete` callback.

Change-Id: I297d2ea93443a73608472f5f682a8b131c74cc76
Bug: 363046220
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6496984
Reviewed-by: Noah Rose Ledesma <noahrose@google.com>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Felipe Pazos <fpazos@google.com>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454261}