Update ScreenAI service termination on idle to ignore clients count.
Previously ScreenAI service was terminated when either all clients got
disconnected, or none of its functionalities were used for 10 seconds.
The threshold was changed to 3s in http://crrev.com/c/6482407 and this
increased the chance of race conditions when service initialization
takes more than 3s and no client is connected to it before
shut down on idle is triggered.

To fix the issue, idle checking is done only based on the last use of
a functionality (initialization is considered a usage). This will delay
service shutdown when all clients are disconnected by at most 3s, but
simplifies the idle checking and prevents crashes due to service shut
down while a client is disconnecting.

As a side change, `ShutDownIfNoClients` interface to the service is
removed and the only usage (in test) relies on normal behavior of the
service.

NO_IFTTT=This is the initial addition for the LINT.IfChange/ThenChange.

AX-Relnotes: n/a
Bug: 413357245, 413922888
Change-Id: I3088c7a4ccd5ace0aafe023ec94ce3ab0e162653
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6494599
Reviewed-by: Kyungjun Lee <kyungjunlee@google.com>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Auto-Submit: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454324}