Rename 'open in browser' to 'open link'.
The functionality of the "open in browser" menu action item is to open
the selected link on the relevant application that can handle the
intent. For example, if the user selects a Youtube link and clicks this
action item, they will be taken to the Youtube app and not to their
default browser. As such, it makes more sense to rename this action item
to "open link". Please see the bug for more details.

This CL also orders the menu items so that "copy link address" comes
first. This is the recommendation of the UX designer.

Bug: 353686611
Change-Id: I319a11ce9842ad886144f608e10814e5690f9478
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6434209
Reviewed-by: Peter Conn <peconn@chromium.org>
Commit-Queue: Zakarie Yaris <yaris@google.com>
Reviewed-by: Alex Mitra <alexmitra@chromium.org>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454388}