Add new Graphite Precompilation PaintOptions
The new set uses the latest collected Pipeline labels. Together they
generate 141 Pipelines of which 113 were in the corpus (yielding an over-generation of 28 Pipelines).

The UMA stats also behave as expected with this set. Compared to a baseline of no web page, loading one of the corpus' web pages yields an increased usage of the Precompiled Pipelines.

Low-Coverage-Reason: EXPERIMENTAL_CODE - not yet enabled in Chrome
Bug: 358074434
Change-Id: I809e3c854e99f28a98c424145c1aafd9b8d1f8dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6500977
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Robert Phillips <robertphillips@google.com>
Cr-Commit-Position: refs/heads/main@{#1454491}