Use GenAiDefaultSettings as default for GeminiSettings
Currently, GeminiSettings does not use GenAiDefaultSettings policy as
a default value when unset. This change updates GeminiSettings to
use GenAiDefaultSettings policy when GeminiSettings is unset. As a
result, when GeminiSettings is unset, `browser.gemini_settings` pref
is set to GenAiDefaultSettings.

GeminiSettings does not support all the enums that GenAiDefaultSettings
supports. As a result, we create a map between GenAiDefaultSettings
value and the value to use for GeminiSettings.
  - GenAiDefaultSettings: 0; GeminiSettings: 0
  - GenAiDefaultSettings: 1; GeminiSettings: 0
  - GenAiDefaultSettings: 2; GeminiSettings: 1

This change adds a translator to
GenAiDefaultSettingsPolicyHandler which can translate the integer value
from GenAiDefaultSettings policy to a specific `int` based on
the `policy_value_to_pref_map`. This change also updates the info message
in chrome://policy to exclude policies using translator map when
feature is disabled.

If no `policy_value_to_pref_map` is set, the GenAI policy will use the
`GenAiDefaultSettings` int value when unset. If
`policy_value_to_pref_map` is set and the `GenAiDefaultSettings`
int value exists in the map, the GenAI policy will use the
value in `policy_value_to_pref_map` when unset. If
`policy_value_to_pref_map` is set and the `GenAiDefaultSettings` int
value does not exist in the map, the GenAI policy will remain unset.
A CHECK has been added to `CheckPolicySettings()` to verify every map
includes all supported GenAiDefaultSettings values. That list should
be updated if new setting values are added.

Mapping is only enabled if `GenAiPolicyDefaultsUsePrefMap` is
enabled. If `GenAiPolicyDefaultsUsePrefMap` is not enabled,
GeminiSettings will use its own default value (0) when unset,
regardless of GenAiDefaultSettings. `GenAiPolicyDefaultsUsePrefMap`
is enabled by default.

Bug: 393097032,413720054
Change-Id: If0c7f78a30681a5979e3e1beb2d8808b8c877b18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6495892
Reviewed-by: Salma Elmahallawy <esalma@google.com>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Alex Chen <alexwchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454526}