Allow embedder to override all binders
There is currently the ability for embedders to override the Binders
registered within PopulateBinderMapWithContext but not those which
are registered in PopulateFrameBinders.  Since the resolution of the
Binders in BrowserInterfaceBrokerImpl resolves those in order it is
impossible to override any interface that is defined in the
BinderMap.

It has been determined that PopulateFrameBinders is legacy and that
PopulateBinderMapWithContext should be preferred everywhere.
Therefore, move the blink::mojom::Authenticator binder from the
PopulateFrameBinders to PopulateBinderMapWithContext and remove the
testing override for it as this can now be performed through the
override mechanism.

Update the single use of this testing override to use the
ChromeContentBrowserClient override mechanism.

Bug: 405869197
Change-Id: I94bba3ae7c9b5a6093552b324222e588e2adbe3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6416947
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Auto-Submit: Christopher Staite <christopher.staite@menlosecurity.com>
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454527}