[Viewport Segments] Fix WPT tests not working on wpt.fyi
It turns out that the runner may have a viewport with dimensions which
will not return an integer if divided by 2 (to put the display feature
in the middle). set_display_features expect an integer so let's make sure we give it one. Also make sure we allow for rounding errors when checking the values of the segments.

Bug: 40137084
Change-Id: I713be16200f78cdca15e80d0397ee930168a727e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6505780
Commit-Queue: Alexis Menard <alexis.menard@intel.com>
Reviewed-by: Jonathan Lee <jonathanjlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1454533}