[ios] Support multiprofile with app notification settings
This change adds support for switching profiles with the "provides app
notification settings" feature. This allows a user to swipe on a
notification, and choose "Options" -> "Configure in Chrome" to open
Chrome. If the notification references a different profile than is
currently active, the profile will be switched before the notification
settings page will be shown.

Bug: 405087334
Change-Id: Icc66dff319701a31bdd8f0d5cea5d6d0ae22128b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6502615
Commit-Queue: Scott Yoder <scottyoder@google.com>
Reviewed-by: Benjamin Williams <bwwilliams@google.com>
Cr-Commit-Position: refs/heads/main@{#1454539}