Send Glic user status RPC only for enterprise accounts
This CL add a check that the Glic user status RPC is sent only for
only enterprise accounts. The checking uses policy::ManagementService
API following https://chromium-review.googlesource.com/c/chromium/src/+/6485427.
The check is done after refresh token is available and before the RPC
is to be sent. This is to allow the chrome policy to be fetched.

Bug: 409647791
Change-Id: Ib164a869dd8d2de98f0629132b4395fb6c5e2646
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6496672
Reviewed-by: Justin DeWitt <dewittj@chromium.org>
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454540}