extensions: Port ExtensionServiceTest to desktop Android
Desktop Android now uses ExtensionService, so we should enable as
many of its unit tests as possible.

Most tests port over cleanly. DesktopAndroidExtensionBrowserClient
AreExtensionsEnabled() is be updated for one test to pass. Several
test support classes are added to the build.

Some tests are skipped because we don't yet support the install and
uninstall error UI, as well as a few other dependencies.

One test is skipped because uninstalling an extension installed from
a zip file doesn't seem to remove the directory. I wasn't able to
figure it out in a reasonable amount of time, so I filed a bug.
Zip file support was added recently, perhaps this is related.

Bug: 413460628, 414911435
Change-Id: Ie4a8983ecba36e7d7ccbc7321a4fca9b479f6ed1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6502239
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454569}