[TreesInViz] Send cc::Tile deletion updates to viz.
Currently when a Tile is destroyed in Renderer, the notification is not
sent to Viz and hence Viz does not delete its copy of those Tiles and
eventually runs out of tile memory limits.

This CL fixes that by calling
LayerTreeHostImpl::NotifyTileStateChanged() when a Tile is being
destroyed. Pending tree tile deletions are sent as a part of
|updated_tiles_| and active tree Tile deletions are sent immediately.

Note that when whole Tiling is deleted, it also deletes all the
corresponding Tiles which will result in
LayerTreeHostImpl::NotifyTileStateChanged(). If the Tiling deletion
happens on Pending Tree, none of the corresponding Tile deletions will
be send to Viz as there is no Tiling for those Tiles anymore in the
PictureLayerTilingSet. This is fine since future CLs will send Tiling
removal notifications to Viz and Viz can simply delete all the Tiles
corresponding to those Tilings.

If Tiling is deleted on active tree, all its corresponding Tile deletion
notifications will be send immediately to Viz irrespective of that
Tiling no longer present in PictureLayerTilingSet. When future CL send
Tiling removal notifications to viz, we should not send all its
corresponding Tile removal notifications.

Bug: 403385455
Change-Id: I5973ee49ce8e9d9e20b5c9dcd7d1105f2a7499e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6491732
Reviewed-by: Victor Miura <vmiura@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Auto-Submit: vikas soni <vikassoni@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454591}