Shut down BrowserTestBase's HTTP test server in TearDown().
BrowserTestBase::embedded_https_test_server_ is only stopped when
BrowserTestBase itself is destroyed. It's not unusual for test fixtures
to bind a raw pointer to themselves to a handler attached to the
EmbeddedTestServer.  As a result, the server will often have a callback
with a raw pointer to the child class.  When the child class is
destroyed, that becomes an invalid pointer, but the EmbeddedTestServer
is only shut down (and destroyed) after that, in BrowserTestBase's
destructor.  The fix is to shut down the server in
BrowserTestBase::TearDown(), like we do for the embedded_test_server_.

It would be safer to delete both test servers there, but I suspect
that may break some tests.

Fixed: 415057190
Change-Id: I4e30746ce40cfb5f650a7471c83c63ac28c6ab33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6506423
Commit-Queue: mmenke <mmenke@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454644}