QuickAnswersBrowserTest: fix paint-holding dependency.
To mitigate a bug with user interactions with a stale page, we are
planning to hold back input events from a page while paint-holding is
active. Paint-holding is the browser feature to continue to show a
snapshot of the pre-navigation page (instead of a blank page) until the
renderer for the post-navigation page has pushed content to the GPU.

The test modified here sends an input event to pages without waiting for
the end of paint-holding. Because the test is not related to
paint-holding, we are adding a forced end of paint-holding w/o affecting
the test's purpose.

Bug: 40057499


Enable for testing.

Change-Id: I0a585d4d2c4336c1a2eb8e80354a6347e575e9d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6500975
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Mustaq Ahmed <mustaq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454649}