[Tab Groups] Fix missing event when merging a tab group to a single tab
When merging a tab group to a tab the event to mark the single tab as
being part of a tab group was not emitted. This led to tab group sync
not recognizing the tab as part of a tab group leading to it missing
from TabGroupSyncService. This has knock on effects to several features
that treat the sync service as a source of truth.

Ungroup was also broken after this fix as didMoveTabOutOfGroup happens
after group information is lost meaning the single tab lost its group
information. This CL fixes that by changing to willMoveTabOutOfGroup
which is emitted in a pair with didMoveTabOutOfGroup before the group
information is destroyed.

Fixed: 414525193
Change-Id: I850cd1871ca8b2f6e90e4de0fd833aaa6ad20f89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6506145
Commit-Queue: Sky Malice <skym@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454671}