[SxS] Loosen coupling between split tab menu and the active tab
Currently split tab icons are determined based on the assumption the
active tab is in a split. This CL determines which icon to show based on
which tab in a split was last active instead of which one is currently
active.

There are also multiple instances of retrieving the active tab from the
tab strip model to get the split id. This CL puts this logic into a
single function to remove code duplication and makes it easier to add
code to support cases when the menu should be shown but the active tab
is not in a split (ex. opening tab context menu on an inactive split
tab).

Bug: 403350993
Change-Id: I9b212c5a51f1f6099a1a08c203ca4247545e829f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6498652
Commit-Queue: Steven Luong <stluong@chromium.org>
Reviewed-by: Alison Gale <agale@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454685}