Allow SecKeychain and SecItem implementations of AppleKeychain to coexist
`AppleKeychain` is converted to an abstract base class with
`AppleKeychainSecKeychain`, `AppleKeychainSecItem`, and
`MockAppleKeychain` as concerete implementations. Both
`AppleKeychainSecKeychain` and `AppleKeychainSecItem` are compiled on
macOS.

A static method `AppleKeychain::DefaultKeychain` is provided to access
the platform's default keychain (file-based keychain on macOS, data
protection keychain on iOS). This returns an instance of
`AppleKeychainSecKeychain` on macOS and `AppleKeychainSecItem` on iOS.
In the future a feature flag will be used to return
`AppleKeychainSecItem` on macOS as well.

`OSCryptImpl::DeriveKey` is modified to always use `KeychainPassword`
for retrieving the password, even when using a mock keychain.

Bug: 395659797
Change-Id: I6a6a21008c8380786ed80e568ffae9c88e41c716
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6444777
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Matt Mueller <mattm@chromium.org>
Commit-Queue: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454709}