Change error type of tabs service to mojo_base's Error
The base lib's Error type provides a set of common codes and the
ability to set a developer readable message. This is probably
good enough for the timebeing.

This change discovered some issues:

  1. Mojo's result return does not properly unwrap the result
     shadow type.
  2. The impl needs to own the receiver somehow, but it also
     needs  to talk to multiple clients, so we should probably
     add a receiver set to the impl.

This change also adds a very simple unit test so that we can
easily keep track of behaviour as we make changes.

Bug: 409821664
Change-Id: I81b026d21742de500fad078f6418219c744142e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6503917
Reviewed-by: David Pennington <dpenning@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: David Yeung <dayeung@chromium.org>
Commit-Queue: Fred Shih <ffred@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454720}