ash: remove DesksTest.DeskTraversalNonTouchpadMetrics
There's a metric that counts the number of desk changes that have been
done during a short interval (5 seconds). The deleted test had code to
verify that the metric is accurately recorded. However, the test
inherently relies on some things completing in a timely fashion and on
bots that doesn't always happen.

More details in the bug.

BUG=b/405887243

Change-Id: Ibd7c4d0045d638ab82e1571883e40764d05f80df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6505802
Reviewed-by: Xiaodan Zhu <zxdan@chromium.org>
Commit-Queue: Daniel Andersson <dandersson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454744}