[PDF Ink Signatures] Add TextHighlightState
Add a new possible tool state for text highlighting in PdfInkModule.
While it is not possible to enter this state in this CL, it will be used
to implement text highlighting in https://crrev.com/c/6313743.

Note that the original invariant in PdfInkModule was that it must either
be in a drawing stroke state or eraser state. This CL alters the
invariant so that it could also be in the text highlighting state. Add
TODOs to update some code spots that rely on the original invariant.

Bug: 342445982
Change-Id: I546bcf49e8b98e3b4bf20cf7b496e5838ef070d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6329567
Reviewed-by: Alan Screen <awscreen@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Andy Phan <andyphan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454748}